First pass at invoicing (#3)
Frontend for invoicing.

Still a work in progress. Allows admin to select jobs that are in "ready to invoice" state
and populates an invoice with those jobs. Allows admin to set a price for each line item,
customer's address, and shipping cost.

Data not written back to database yet, this is just the frontend.